token endpoint ready

+++ app/api/serializer/products.py
@@ -5,7 +5,7 @@ product_serializer = {
     "id": fields.Integer,
     "sku": fields.String,
     "name": fields.String,
-    "price": fields.Decimal,
+    "price": fields.Float,
     "brand": fields.String,
     "description": fields.String,
     "qty": fields.Integer,
